+++ src/main/java/com/dazzhub/skywars/Utils/achievements/IAchievement.java
@@ -96,7 +96,7 @@ public class IAchievement {
                         coins = coins.substring(1);
                     }
 
-                    gamePlayer.addCoins(Integer.parseInt(coins));
+                    gamePlayer.addCoins(Double.parseDouble(coins));
                 } else if (cmd.startsWith("cage:")) {
                     String cage = cmd.substring(5);
                     if (cage.startsWith(" ")) {
